+++ metrics/prowlarr/queries.test.ts
@@ -1,0 +1,5 @@
+import { expectFixtureMetric } from '../test-utils'
+
+it('sums Prowlarr queries', async () => {
+  await expectFixtureMetric(new URL('./queries.yml', import.meta.url), { indexers: [{ numberOfQueries: 11 }, { numberOfQueries: 13 }] }, 24)
+})